Updated beer discount calculation and removed uneeded files. Also added float prices

+++ groceries/Backend/src/main/java/dev/yashkalia/groceries/Models/Items/Item.java
@@ -13,7 +13,7 @@ public class Item {
 
 //    @Id
     private String ItemName;
-    private Integer Price;
+    private Float Price;
     private String ItemType;
     private String id;
     private Integer Age;
